Extracted Keypair
