+++ rosenpass/tests/api-integration-tests-api-setup.rs
@@ -74,7 +74,7 @@ fn api_integration_api_setup(protocol_version: ProtocolVersion) -> anyhow::Resul
     let peer_a_endpoint = "[::1]:0";
     let peer_a_listen = std::net::UdpSocket::bind(peer_a_endpoint)?;
     let peer_a_endpoint = format!("{}", peer_a_listen.local_addr()?);
-    let peer_a_keypair = config::Keypair::new(tempfile!("a.pk"), tempfile!("a.sk"));
+    let peer_a_keypair = config::rp_keypair::Keypair::new(tempfile!("a.pk"), tempfile!("a.sk"));
 
     let peer_b_osk = tempfile!("b.osk");
     let peer_b_wg_device = "mock_device";
@@ -101,7 +101,7 @@ fn api_integration_api_setup(protocol_version: ProtocolVersion) -> anyhow::Resul
             key_out: None,
             endpoint: None,
             pre_shared_key: None,
-            wg: Some(config::WireGuard {
+            wg: Some(config::wireguard::WireGuard {
                 device: peer_b_wg_device.to_string(),
                 peer: format!("{}", peer_b_wg_peer_id.fmt_b64::<8129>()),
                 extra_params: vec![],
@@ -111,7 +111,7 @@ fn api_integration_api_setup(protocol_version: ProtocolVersion) -> anyhow::Resul
         }],
     };
 
-    let peer_b_keypair = config::Keypair::new(tempfile!("b.pk"), tempfile!("b.sk"));
+    let peer_b_keypair = config::rp_keypair::Keypair::new(tempfile!("b.pk"), tempfile!("b.sk"));
     let peer_b = config::Rosenpass {
         config_file_path: tempfile!("b.config"),
         keypair: Some(peer_b_keypair.clone()),
