+++ rosenpass/tests/config_Rosenpass_parse_args_simple.rs
@@ -3,7 +3,7 @@ use std::{
     path::PathBuf,
 };
 
-use rosenpass::config::{Keypair, Rosenpass, RosenpassPeer, Verbosity};
+use rosenpass::config::{rp_keypair::Keypair, Rosenpass, RosenpassPeer, Verbosity};
 
 #[test]
 fn parse_simple() {
